Module spotify::status

The status module.

This module contains methods to turn the JSON responses from the Spotify connector into easy-to-use structures.

It also contains some extra abstractions, such as the SimpleTrack struct.

Structs

OpenGraphState
A Spotify Open Graph state.

Resource
A Spotify resource.

ResourceLocation
A Spotify resource location.

SimpleTrack
A simple track. Provides an abstraction over the more complicated and quite messy Track struct.

SpotifyStatus
A Spotify status.

Track
A Spotify track.